Get project to build; cut new AMIs
The `m3.large` EC2 reservation for the instance that runs the GeoTrellis
documentation site runs out in September. In order to renew the
reservation at another instance type (hopefully, `t2.medium`), I had to
get the project building again, and generate a new AMI of the working
setup.

Most of the changes under `deployment/` were to reorganize things like
our other AMI based projects. All other changes are related to getting
the project building again, as a bunch of the dependencies have changed.

Thanks to @echeipesh for helping me get through this.